requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.18
openai>=1.3.0
pypdf2>=3.0.1
python-dotenv>=1.0.0
tiktoken>=0.5.2

# PDF processing
pymupdf>=1.23.0
fitz>=0.0.1.dev2

# Vector store and embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Web interface
streamlit-chat>=0.1.1
streamlit-option-menu>=0.3.6

# Utilities
pydantic>=2.5.0
pandas>=2.1.0
numpy>=1.24.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0

# Optional: FastAPI backend
fastapi>=0.104.0
uvicorn>=0.24.0
